--HG--
extra : convert_revision : svn%3A83215879-3e5a-4751-8c9d-778f44bb06a5/trunk%40300

+++ testing/thrusttest/testframework.h
@@ -185,7 +185,7 @@ template<template <typename> class TestName, typename TypeList>
             1024, 1025, 10000, 12345,                     
             (1<<16) - 3, 1<<16, (1<<16) + 3};             
         size_t num_sizes = sizeof(sizes) / sizeof(size_t);
-        for(int i = 0; i != num_sizes; ++i)               
+        for(size_t i = 0; i != num_sizes; ++i)               
         {                                                 
             // get the first type in the list
             typedef typename thrusttest::get_type<TypeList,0>::type first_type;
